refactor(useProducts.ts): useProducts 커스텀 훅을 역할에 따라 useCartItems, useProductFetch, useProductSelection으로 분리
